# enterprise_openapi.json
{"openapi":"3.0.1","info":{"title":"OpenAPI definition","version":"v0"},"servers":[{"url":"http://localhost:8080/","description":"Generated server url"}],"paths":{"/enterprise":{"get":{"tags":["enterprise-controller"],"operationId":"getEnterprises","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Enterprise"}}}}}}},"put":{"tags":["enterprise-controller"],"operationId":"saveOrUpdateEnterprise","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Enterprise"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Enterprise"}}}}}},"post":{"tags":["enterprise-controller"],"operationId":"addEnterprise","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Enterprise"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Enterprise"}}}}}}},"/enterprise/{eid}":{"get":{"tags":["enterprise-controller"],"operationId":"getEnterprise_1","parameters":[{"name":"eid","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Enterprise"}}}}}},"put":{"tags":["enterprise-controller"],"operationId":"getEnterprise_5","parameters":[{"name":"eid","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Enterprise"}}}}}},"post":{"tags":["enterprise-controller"],"operationId":"getEnterprise_4","parameters":[{"name":"eid","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Enterprise"}}}}}},"delete":{"tags":["enterprise-controller"],"operationId":"deleteEnterprise_1","parameters":[{"name":"eid","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"oneOf":[{"$ref":"#/components/schemas/Enterprise"},{"type":"string"}]}}}}}},"options":{"tags":["enterprise-controller"],"operationId":"getEnterprise_3","parameters":[{"name":"eid","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Enterprise"}}}}}},"head":{"tags":["enterprise-controller"],"operationId":"getEnterprise","parameters":[{"name":"eid","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Enterprise"}}}}}},"patch":{"tags":["enterprise-controller"],"operationId":"getEnterprise_2","parameters":[{"name":"eid","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Enterprise"}}}}}}},"/healthcheck":{"get":{"tags":["enterprise-controller"],"operationId":"getHealth_1","parameters":[{"name":"eid","in":"query","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"string"}}}}}},"put":{"tags":["enterprise-controller"],"operationId":"getHealth_6","parameters":[{"name":"eid","in":"query","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"string"}}}}}},"post":{"tags":["enterprise-controller"],"operationId":"getHealth_5","parameters":[{"name":"eid","in":"query","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"string"}}}}}},"delete":{"tags":["enterprise-controller"],"operationId":"getHealth_2","parameters":[{"name":"eid","in":"query","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"string"}}}}}},"options":{"tags":["enterprise-controller"],"operationId":"getHealth_4","parameters":[{"name":"eid","in":"query","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"string"}}}}}},"head":{"tags":["enterprise-controller"],"operationId":"getHealth","parameters":[{"name":"eid","in":"query","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"string"}}}}}},"patch":{"tags":["enterprise-controller"],"operationId":"getHealth_3","parameters":[{"name":"eid","in":"query","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"string"}}}}}}},"/":{"get":{"tags":["enterprise-controller"],"operationId":"home_1","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"string"}}}}}},"put":{"tags":["enterprise-controller"],"operationId":"home_6","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"string"}}}}}},"post":{"tags":["enterprise-controller"],"operationId":"home_5","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"string"}}}}}},"delete":{"tags":["enterprise-controller"],"operationId":"home_2","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"string"}}}}}},"options":{"tags":["enterprise-controller"],"operationId":"home_4","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"string"}}}}}},"head":{"tags":["enterprise-controller"],"operationId":"home","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"string"}}}}}},"patch":{"tags":["enterprise-controller"],"operationId":"home_3","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"string"}}}}}}}},"components":{"schemas":{"Enterprise":{"type":"object","properties":{"eid":{"type":"integer","format":"int64"},"name":{"type":"string"},"url":{"type":"string"},"authentication":{"type":"string"},"health":{"type":"boolean"}}}}}}